backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-tutor-backend"
version = "0.1.0"
description = "Math Tutor FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "openai>=1.51.0",
  "pdfplumber>=0.11.0",
  "pymupdf>=1.24.0",
  "pytesseract>=0.3.10",
  "pillow>=10.0.0",
  "pydantic>=2.8.0",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.1",
  "redis>=5.0.0",
  "rq>=1.16.0",
  "sqlalchemy>=2.0.32",
  "uvicorn[standard]>=0.30.0",
  "psycopg2-binary>=2.9.9",
  "faster-whisper>=1.0.3",
  "manim>=0.18.1",
  "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "pytest-cov>=5.0.0",
  "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
